requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sentence-transformers>=3.0.1
endee==0.1.6
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.1
torch>=2.9.0
transformers>=4.41.2
numpy>=2.2.4
pandas>=2.2.2
tqdm>=4.66.4
